Move modal component to components/layout
