+++ src/components/layout/Layout.module.scss
@@ -70,3 +70,24 @@
     }
   }
 }
+
+.modal {
+  &__container {
+    position: fixed;
+    width: 100%;
+    height: 100vh;
+    bottom: 0;
+    z-index: 1000;
+    overflow-y: auto;
+    background: white;
+    border: none;
+    border-top-left-radius: 16px;
+    border-top-right-radius: 16px;
+    padding: 40px;
+    box-shadow: var(--shadow-cart);
+    @include media($breakpoint-sm) {
+      border-radius: 16px;
+      box-shadow: var(--shadow-lg);
+    }
+  }
+}
